requirements.txt
transformers
datasets
torch
scikit-learn
pandas
numpy
streamlit
python-docx
PyPDF2
